requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
ultralytics>=8.0.0
PyYAML>=6.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
requests>=2.27.0
python-dateutil>=2.8.2
cryptography>=41.0.0
jupyter>=1.0.0
ipython>=8.0.0
